+++ spec/legion/extensions/telemetry/actors/region_reporter_spec.rb
@@ -17,16 +17,16 @@ require 'legion/extensions/telemetry/actors/region_reporter'
 RSpec.describe Legion::Extensions::Telemetry::Actor::RegionReporter do
   subject(:actor) { described_class.new }
 
-  it 'runs every 120 seconds' do
-    expect(actor.time).to eq(120)
+  it 'runs every 60 seconds' do
+    expect(actor.time).to eq(60)
   end
 
   it 'targets the telemetry runner' do
     expect(actor.runner_class).to eq('Legion::Extensions::Telemetry::Runners::Telemetry')
   end
 
-  it 'calls region_metrics' do
-    expect(actor.runner_function).to eq('region_metrics')
+  it 'calls region_stats' do
+    expect(actor.runner_function).to eq('region_stats')
   end
 
   it 'does not run immediately' do